+++ kernel/include/idt.h
@@ -58,4 +58,4 @@ extern void isr31();
 
 void init_idt();
 
-void idt_set_gate(uint8_t num, uint32_t base, uint16_t sel, uint8_t flags);
+void idt_set_gate(int num, uint32_t base, uint16_t sel, uint8_t flags);
